check if file exists
Former-commit-id: 64fee143ad2f3ea082df988486a5d5c78b3aac6a

+++ script/remove_old_recordings.rb
@@ -2,6 +2,8 @@ old_recordings = Recording.where("created_at < ?", 7.days.ago)
 puts "Removing #{old_recordings.count} old recordings..."
 
 old_recordings.find_each do |recording|
-  File.delete recording.path
+  if File.exist? recording.path
+    File.delete recording.path
+  end
   recording.destroy
 end
